fix bug preventing first-sentence searching and add style
an oversight in the RegExp constructor prevented matching anything until after the first sentence-ending punctuation mark ('.', '?', or '!'). specifically, the '(?<=\\.|\\!|\\?)\\s' part was the issue. this bug has now been fixed.

added styling to the output beyond just the basic highlighting in oput2. now the specific keyword detected in any section printing a match will be underlined.

general nice stuff:
made things more modular. dunno if or when, but i have some ideas i may implement (hyperlinks, sorting, etc.), and making it more modular makes it easier to a) fix ^^^ bugs and b) work with this in the future.